Split MediaViewModel into multiple files
